+++ gems/aws-sdk-core/lib/seahorse/client/net_http/connection_pool.rb
@@ -35,8 +35,8 @@ module Seahorse
           ssl_ca_directory: nil,
           ssl_ca_store: nil,
           ssl_timeout: nil,
-          http_client_cert: nil,
-          http_client_key: nil
+          ssl_cert: nil,
+          ssl_key: nil
         }
 
         # @api private
@@ -249,8 +249,8 @@ module Seahorse
               :ssl_ca_directory => options[:ssl_ca_directory],
               :ssl_ca_store => options[:ssl_ca_store],
               :ssl_timeout => options[:ssl_timeout],
-              :http_client_cert => options[:http_client_cert],
-              :http_client_key => options[:http_client_key]
+              :ssl_cert => options[:ssl_cert],
+              :ssl_key => options[:ssl_key]
             }
           end
 
@@ -295,8 +295,8 @@ module Seahorse
               http.ca_file = ssl_ca_bundle if ssl_ca_bundle
               http.ca_path = ssl_ca_directory if ssl_ca_directory
               http.cert_store = ssl_ca_store if ssl_ca_store
-              http.cert = http_client_cert if http_client_cert
-              http.key = http_client_key if http_client_key
+              http.cert = ssl_cert if ssl_cert
+              http.key = ssl_key if ssl_key
             else
               http.verify_mode = OpenSSL::SSL::VERIFY_NONE
             end