+++ gems/aws-sdk-core/lib/seahorse/client/plugins/net_http.rb
@@ -70,12 +70,12 @@ Sets the X509::Store to verify peer certificate.
           resolve_ssl_timeout(cfg)
         end
 
-        option(:http_client_cert, default: nil, doc_type: OpenSSL::X509::Certificate, docstring: <<-DOCS)
+        option(:ssl_cert, default: nil, doc_type: OpenSSL::X509::Certificate, docstring: <<-DOCS)
 Sets a client certificate when creating http connections.
         DOCS
 
 
-        option(:http_client_key, default: nil, doc_type: OpenSSL::PKey, docstring: <<-DOCS)
+        option(:ssl_key, default: nil, doc_type: OpenSSL::PKey, docstring: <<-DOCS)
 Sets a client key when creating http connections.
         DOCS