+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased Changes
 ------------------
 
-* Issue - Add support for Http client cert and key to support mTLS.
+* Issue - Add support for `ssl_cert` and `ssl_key` configuration options to support mTLS.
 
 3.203.0 (2024-09-03)
 ------------------